refactor: extract junit report parsing code in JunitReportReader
