+++ application/src/main/kotlin/application/mcp/server/tools/ContractTestTool.kt
@@ -1,11 +1,9 @@
 package application.mcp.server.tools
 
+import application.mcp.utils.JunitReportReader
 import application.TestCommand
 import io.specmatic.core.utilities.Flags.Companion.SPECMATIC_GENERATIVE_TESTS
-import io.specmatic.core.utilities.newXMLBuilder
 import kotlinx.serialization.Serializable
-import org.w3c.dom.Element
-import org.w3c.dom.Node
 import picocli.CommandLine
 import java.io.File
 import kotlin.io.path.createTempDirectory
@@ -45,7 +43,7 @@ class ContractTestTool(
                 CommandLine(command).execute(*argsList.toTypedArray())
             }
 
-            val summary = parseJUnitSummary(junitReportFile)
+            val summary = JunitReportReader().parseJUnitSummary(junitReportFile)
 
             formatTestResult(
                 title = if (resiliency) "Resiliency" else "Contract",
@@ -65,37 +63,7 @@ class ContractTestTool(
         }
     }
 
-    internal fun parseJUnitSummary(reportFile: File): TestSummary? {
-        if (!reportFile.isFile) return null
 
-        val root = newXMLBuilder().parse(reportFile).documentElement ?: return null
-        if (root.tagName !in setOf("testsuite", "testsuites")) return null
-
-        val testCases = root.descendantElements("testcase")
-        val failedTests = testCases.mapNotNull { testCase ->
-            val failure = testCase.firstChildElement("failure") ?: testCase.firstChildElement("error") ?: return@mapNotNull null
-            FailedTest(
-                scenario = testCase.getAttribute("name"),
-                message = failure.getAttribute("message").ifBlank { failure.textContent.orEmpty().trim() }
-            )
-        }
-
-        val total = root.intAttribute("tests") ?: testCases.size
-        val failures = root.intAttribute("failures")
-        val errors = root.intAttribute("errors")
-        val failed = if (failures != null || errors != null) (failures ?: 0) + (errors ?: 0) else failedTests.size
-        val skipped = root.intAttribute("skipped") ?: testCases.count { it.firstChildElement("skipped") != null }
-        val passed = total - failed - skipped
-
-        return TestSummary(
-            total = total,
-            passed = passed,
-            failed = failed,
-            failedTests = failedTests,
-            reportLabel = "JUnit",
-            reportPath = reportFile.canonicalPath
-        )
-    }
 
     private fun formatTestResult(
         title: String,
@@ -167,22 +135,7 @@ private fun createMcpJUnitReportDirectory(resiliency: Boolean): File {
     return createTempDirectory(baseDir.toPath(), "$prefix-").toFile()
 }
 
-private fun Element.intAttribute(name: String): Int? {
-    return getAttribute(name).takeIf { it.isNotBlank() }?.toIntOrNull()
-}
-
-private fun Element.descendantElements(tagName: String): List<Element> {
-    val nodes = getElementsByTagName(tagName)
-    return (0 until nodes.length).mapNotNull { nodes.item(it) as? Element }
-}
-
-private fun Element.firstChildElement(tagName: String): Element? {
-    return (0 until childNodes.length)
-        .map { childNodes.item(it) }
-        .firstOrNull { it.nodeType == Node.ELEMENT_NODE && it.nodeName == tagName } as? Element
-}
-
-internal data class TestSummary(
+data class TestSummary(
     val total: Int,
     val passed: Int,
     val failed: Int,
@@ -191,7 +144,7 @@ internal data class TestSummary(
     val reportPath: String
 )
 
-internal data class FailedTest(
+data class FailedTest(
     val scenario: String,
     val message: String
 )
