+++ core/src/main/kotlin/io/specmatic/core/config/LoggingConfiguration.kt
@@ -40,13 +40,8 @@ data class LogOutputConfig(val directory: String?, val console: Boolean?, @field
     }
 
     companion object {
-        fun defaultOr(logOutputConfig: LogOutputConfig?): LogOutputConfig {
-            val defaultConfig = LogOutputConfig(directory = null, console = true)
-            return when {
-                logOutputConfig == null -> defaultConfig
-                logOutputConfig.directory == null && logOutputConfig.console == null -> defaultConfig
-                else -> logOutputConfig
-            }
+        fun default(): LogOutputConfig {
+            return LogOutputConfig(directory = null, console = true)
         }
     }
 }
@@ -59,10 +54,10 @@ data class LoggingConfiguration(
     fun levelOrDefault(): ConfigLoggingLevel = level ?: ConfigLoggingLevel.INFO
 
     fun hasJsonConfiguration(): Boolean = json != null
-    fun jsonConfigurationOrDefault(): LogOutputConfig = LogOutputConfig.defaultOr(json)
+    fun jsonConfigurationOrDefault(): LogOutputConfig = json ?: LogOutputConfig.default()
 
     fun hasTextConfiguration(): Boolean = text != null
-    fun textConfigurationOrDefault(): LogOutputConfig = LogOutputConfig.defaultOr(text)
+    fun textConfigurationOrDefault(): LogOutputConfig = text ?: LogOutputConfig.default()
 
     fun overrideMergeWith(other: LoggingConfiguration?): LoggingConfiguration {
         if (other == null) return this
@@ -81,8 +76,17 @@ data class LoggingConfiguration(
         fun from(data: LoggingFromOpts): LoggingConfiguration {
             return LoggingConfiguration(
                 level = if (data.debug == true) ConfigLoggingLevel.DEBUG else null,
-                text = LogOutputConfig(directory = data.textLogDirectory?.path, console = data.textConsoleLog, logFilePrefix = data.logPrefix),
-                json = LogOutputConfig(directory = data.jsonLogDirectory?.path, console = data.jsonConsoleLog, logFilePrefix = data.logPrefix)
+                text = if(data.textConsoleLog != null){
+                    LogOutputConfig(directory = data.textLogDirectory?.path, console = data.textConsoleLog, logFilePrefix = data.logPrefix)
+                }
+                else{
+                    null
+                },
+                json = if (data.jsonConsoleLog != null && data.jsonLogDirectory != null) {
+                    LogOutputConfig(directory = data.jsonLogDirectory.path, console = data.jsonConsoleLog, logFilePrefix = data.logPrefix)
+                } else {
+                    null
+                },
             )
         }
